Judge a size before converting it, on every int width
The 386 and arm lanes added in #72 failed, and one of the two failures is a
crash reachable on 64-bit as well.

affordDecoded refused a codestream by multiplying its two sides in an int:

	if cw > maxImagePixels || ch > maxImagePixels || cw*ch > maxImagePixels

A codestream declaring 65 535 by 65 535 makes 4 294 836 225. Neither side alone
passes the ceiling, so only the product can refuse it -- and in an int32 that
product wraps to MINUS 131 071, which is not greater than anything. The guard
let the bytes through and the decoder asked for four gigabytes, on a machine
where that allocation cannot even be expressed: image.NewGray panicked. A 1 KB
file crashed a 32-bit build. Its sibling afford in images.go had already been
written in int64 and says why in a comment; this one had not.

decodeBase multiplied the same way and is fixed the same way.

Page was worse, and not only on 32-bit. It converted first and judged after:

	w := int(math.Ceil((box[2] - box[0]) * s))
	...
	if w*h > opt.maxPixels()

A MediaBox is a number in somebody's file and the scale is the caller's, so
that product is an arbitrary float. Converting a float too large for an int is
NOT DEFINED in Go, so the overflow happened before the ceiling could refuse
anything. A page with a MediaBox of 1e300 panics in fill() on arm64 today. The
extent is now judged as a float, before it becomes an int.

TestAPageTooLargeToCountIsRefusedNotConverted covers that one everywhere: it
panics against the old arithmetic on 64-bit, so it is a real regression test
rather than a lane-only one. The codestream defect cannot be reproduced on a
64-bit machine at all -- the 32-bit lane is its only witness, which is the
argument for having added it.

Also in here: TestAPageMayBeGivenOnlySoLong bounded the time budget in one
machine's seconds. The deadline is looked at once every timeCheckEvery
operations, so the overshoot is a quantity in OPERATIONS; five fixed seconds
was that quantity at the speed of the machine that wrote the test, and riscv64,
s390x and 386 took 5.4 s, 8.3 s and 11.3 s for the same correct code. It now
draws the same page twice, once unbounded, and requires the bounded draw to
come back in under a third of it -- measured here it comes back in a ninth.
Everything fixed about the page is paid by both runs and cancels.

Verified in both directions: with timeCheckEvery raised to 200 000 the test
fails; at 256 it passes. Coverage stays at 100.0%.

+++ image.go
@@ -129,7 +129,9 @@ func (r *renderer) decodeImage(dict reader.Dict, raw []byte, resources reader.Di
 func (r *renderer) decodeBase(dict reader.Dict, raw []byte, resources reader.Dict) *sampled {
 	w := int(intOr(resolve(r.doc, dict.Get("Width")), 0))
 	h := int(intOr(resolve(r.doc, dict.Get("Height")), 0))
-	if w <= 0 || h <= 0 || w*h > maxImagePixels {
+	// int64 for the same reason as in affordDecoded: these two numbers come
+	// out of the file, and their product does not fit a 32-bit int.
+	if w <= 0 || h <= 0 || int64(w)*int64(h) > maxImagePixels {
 		return nil
 	}
 	// An image whose filter chain broke part way gives the rows it managed,
@@ -530,22 +532,30 @@ func (r *renderer) affordDecoded(cw, ch, charged int) bool {
 		// the body either: the decoder gives up before it allocates.
 		return true
 	}
-	if cw > maxImagePixels || ch > maxImagePixels || cw*ch > maxImagePixels {
+	// In int64, because int is 32 bits on a 32-bit build and this product is
+	// two numbers out of a file: a codestream declaring 65 535 by 65 535 makes
+	// 4 294 836 225, which wraps to MINUS 131 071 in an int32 and walks
+	// straight through a ceiling written as `>`. Its sibling afford already
+	// said why, one file over. Either side alone is still refused first, so a
+	// width past the ceiling never reaches the multiplication at all.
+	if cw > maxImagePixels || ch > maxImagePixels || int64(cw)*int64(ch) > maxImagePixels {
 		return false
 	}
 	if !r.bounded {
 		return true
 	}
-	extra := cw*ch - charged
+	extra := int64(cw)*int64(ch) - int64(charged)
 	if extra <= 0 {
 		return true
 	}
-	if extra > r.budget {
+	if extra > int64(r.budget) {
 		r.refused = fmt.Errorf("%w: a picture whose codestream holds %d by %d pixels, with %d of the %d pixels left",
 			ErrTooMuchToDecode, cw, ch, r.budget, maxImagesPixels)
 		return false
 	}
-	r.budget -= extra
+	// extra is at most the budget here, and the budget is at most
+	// maxImagesPixels, so this fits an int on every build.
+	r.budget -= int(extra)
 	return true
 }
